修改图片删除

+++ service/news/news.js
@@ -1,6 +1,7 @@
 let db = require('../../dao/mysqlDB');
 let path = require("path");
 let utils = require('../../utils/utile')
+let fs = require('fs')
 /**得到所有新闻类型 */
 exports.getAllNewsTypes = (req, res) => {
     let sql = "SELECT * FROM tb_news_type"
@@ -36,6 +37,7 @@ exports.getAllNews = (req, res) => {
 }
 //根据newsID拿到新闻信息
 exports.getNewsById = (req, res) => {
+    //缓存图片路径
     let sql = "SELECT n.*,typeName FROM tb_news n,tb_news_type t WHERE n.type = t.newsType_id and news_id = ?"
     let sqlParams = [req.body.newID]
     db.query(sql, sqlParams, function (err, result) {
@@ -45,24 +47,39 @@ exports.getNewsById = (req, res) => {
 }
 //根据newsID删除新闻信息
 exports.deleteNewsById = (req, res) => {
-    let sql = "DELETE FROM tb_news WHERE news_id = ?"
-    let sqlParams = [req.body.newID]
-    db.query(sql, sqlParams, function (err, result) {
-        if (err) return res.send({ status: "3306", massage: err.message });
-        if (result.affectedRows)
-            res.send({ status: "0", massage: "删除成功" });
-        else {
-            res.send({ status: "1", massage: "删除失败" });
+    let sqla = 'SELECT pic FROM tb_news WHERE news_id = ?';
+    let sqlParamsa = [req.body.newID];
+    db.query(sqla, sqlParamsa, function (err, resulta) {
+        if (err) throw err;
+        if (resulta) {
+            let sql = "DELETE FROM tb_news WHERE news_id = ?"
+            let sqlParams = [req.body.newID]
+            db.query(sql, sqlParams, function (err, result) {
+                if (err) return res.send({ status: "3306", massage: err.message });
+                console.log(result);
+                if (result.affectedRows){
+                    try {
+                        fs.unlinkSync(path.join(__dirname,'../..'+resulta[0].pic));
+                    } catch (error) {
+                        console.log("图片没有");
+                    }
+                    res.send({ status: "0", massage: "删除成功" });
+                }
+                else {
+                    res.send({ status: "1", massage: "删除失败" });
+                }
+            })
         }
     })
+
 }
 //根据newsID更新新闻信息
 exports.updateNewsById = (req, res) => {
-    let sql='';
-    let sqlParams=[];
-    if(req.file){
-            sql = "UPDATE tb_news SET title=?,author=?,titleDesc=?,content=?, pic=?,update_time= now(),type=?,count=?,comment=? WHERE news_id=?"
-            sqlParams = [
+    let sql = '';
+    let sqlParams = [];
+    if (req.file) {
+        sql = "UPDATE tb_news SET title=?,author=?,titleDesc=?,content=?, pic=?,update_time= now(),type=?,count=?,comment=? WHERE news_id=?"
+        sqlParams = [
             req.body.title,
             req.body.author,
             req.body.titleDesc,
@@ -73,9 +90,9 @@ exports.updateNewsById = (req, res) => {
             0,
             0,
             req.body.newID];
-    }else{
-            sql = "UPDATE tb_news SET title=?,author=?,titleDesc=?,content=?,update_time= now(),type=?,count=?,comment=? WHERE news_id=?"
-            sqlParams = [
+    } else {
+        sql = "UPDATE tb_news SET title=?,author=?,titleDesc=?,content=?,update_time= now(),type=?,count=?,comment=? WHERE news_id=?"
+        sqlParams = [
             req.body.title,
             req.body.author,
             req.body.titleDesc,
@@ -116,14 +133,26 @@ exports.insertNews = (req, res) => {
     console.log(sqlParams);
     db.query(sql, sqlParams, function (err, result) {
         if (err) return res.send({ status: "3306", massage: err.message });
-        if (result.affectedRows||result[0].affectedRows)
+        if (result.affectedRows || result[0].affectedRows)
             res.send({ status: "0", massage: "插入成功" });
         else {
             res.send({ status: "1", massage: "插入失败" });
         }
     })
 }
-
+// 缓存图片路径
+function getUrl(id) {
+    console.log("11");
+    let sql = 'SELECT pic FROM tb_news WHERE news_id = ?';
+    let sqlParams = [id];
+    let value = '';
+    db.query(sql, sqlParams, function (err, result) {
+        if (err) throw err;
+        value = result;
+    })
+    console.log(value);
+    return value;
+}
 // exports.all =(req,res)=>{
 //     let sql = "UPDATE tb_user SET disabled = ? WHERE username =?"
 //     let sqlParams =[]
